requirements.txt
python-dotenv==1.1.1
python-telegram-bot==13.15
requests==2.32.4
urllib3==1.26.15
